separating urlmaking to a different component
