+++ package.json
@@ -7,6 +7,7 @@
     "@material-ui/icons": "^4.2.1",
     "axios": "^0.19.0",
     "eslint-plugin-react-hooks": "^2.0.1",
+    "jalali-moment": "^3.3.3",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "react-persian-calendar-date-picker": "^1.1.5",
